requirements.txt
PyMuPDF
gmft
sentence-transformers
mistralai
langchain
langchain-community
langchain_experimental
langchain-huggingface
tiktoken
ftfy
python-pptx
python-docx
pymilvus
requests
tqdm
tf-keras
python-dotenv
streamlit